+++ Tests/SwashbuckleSettings.All/SwashbuckleSettings.All.Api/Controllers/DefaultController.cs
@@ -1,0 +1,105 @@
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using Intent.RoslynWeaver.Attributes;
+using MediatR;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.DependencyInjection;
+using SwashbuckleSettings.All.Application.CreateWidget;
+using SwashbuckleSettings.All.Application.ReplaceWidget;
+using SwashbuckleSettings.All.Application.UpdateWidget;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: DefaultIntentManaged(Mode.Fully, Targets = Targets.Usings)]
+[assembly: IntentTemplate("Intent.AspNetCore.Controllers.Controller", Version = "1.0")]
+
+namespace SwashbuckleSettings.All.Api.Controllers
+{
+    [ApiController]
+    public class DefaultController : ControllerBase
+    {
+        private readonly ISender _mediator;
+
+        public DefaultController(ISender mediator)
+        {
+            _mediator = mediator ?? throw new ArgumentNullException(nameof(mediator));
+        }
+
+        /// <summary>
+        /// </summary>
+        /// <response code="201">Successfully created.</response>
+        /// <response code="400">One or more validation errors have occurred.</response>
+        [HttpPost("api/widgets")]
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
+        public async Task<ActionResult> CreateWidget(
+            [FromBody] CreateWidget command,
+            CancellationToken cancellationToken = default)
+        {
+            await _mediator.Send(command, cancellationToken);
+            return Created(string.Empty, null);
+        }
+
+        /// <summary>
+        /// </summary>
+        /// <response code="204">Successfully updated.</response>
+        /// <response code="400">One or more validation errors have occurred.</response>
+        /// <response code="404">One or more entities could not be found with the provided parameters.</response>
+        [HttpPut("api/widgets/{widgetId}/replace")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
+        public async Task<ActionResult> ReplaceWidget(
+            [FromRoute] Guid widgetId,
+            [FromBody] ReplaceWidget command,
+            CancellationToken cancellationToken = default)
+        {
+            if (command.WidgetId == Guid.Empty)
+            {
+                command.WidgetId = widgetId;
+            }
+
+            if (widgetId != command.WidgetId)
+            {
+                return BadRequest();
+            }
+
+            await _mediator.Send(command, cancellationToken);
+            return NoContent();
+        }
+
+        /// <summary>
+        /// </summary>
+        /// <response code="204">Successfully updated.</response>
+        /// <response code="400">One or more validation errors have occurred.</response>
+        /// <response code="404">One or more entities could not be found with the provided parameters.</response>
+        [HttpPut("api/widgets/{widgetId}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
+        public async Task<ActionResult> UpdateWidget(
+            [FromRoute] Guid widgetId,
+            [FromBody] UpdateWidget command,
+            CancellationToken cancellationToken = default)
+        {
+            if (command.WidgetId == Guid.Empty)
+            {
+                command.WidgetId = widgetId;
+            }
+
+            if (widgetId != command.WidgetId)
+            {
+                return BadRequest();
+            }
+
+            await _mediator.Send(command, cancellationToken);
+            return NoContent();
+        }
+    }
+}